+++ mitra_models/src/posts/queries.rs
@@ -27,6 +27,7 @@ use crate::profiles::{
 use crate::relationships::types::RelationshipType;
 
 use super::types::{
+    DbEmojiReactions,
     DbPost,
     Post,
     PostCreateData,
@@ -311,6 +312,7 @@ pub async fn create_post(
         db_tags,
         db_links,
         db_emojis,
+        vec![],
     )?;
     transaction.commit().await?;
     Ok(post)
@@ -415,6 +417,22 @@ pub async fn update_post(
         post_data.emojis,
     ).await?;
 
+    // Get emoji reactions
+    let statement = format!(
+        "
+        SELECT {related_emoji_reactions}
+        FROM post
+        WHERE post.id = $1
+        ",
+        related_emoji_reactions=RELATED_EMOJI_REACTIONS,
+    );
+    let emoji_reactions_row = transaction.query_one(
+        &statement,
+        &[&post_id],
+    ).await?;
+    let db_emoji_reactions: Vec<DbEmojiReactions> =
+        emoji_reactions_row.try_get("emoji_reactions")?;
+
     // Create notifications
     for profile in db_mentions.iter() {
         if profile.is_local() &&
@@ -440,6 +458,7 @@ pub async fn update_post(
         db_tags,
         db_links,
         db_emojis,
+        db_emoji_reactions,
     )?;
     transaction.commit().await?;
     let deletion_queue = DeletionQueue {
@@ -484,6 +503,23 @@ pub(crate) const RELATED_EMOJIS: &str =
         JOIN emoji ON post_emoji.emoji_id = emoji.id
         WHERE post_emoji.post_id = post.id
     ) AS emojis";
+
+pub(crate) const RELATED_EMOJI_REACTIONS: &str =
+    "ARRAY(
+        SELECT
+            json_build_object(
+                'authors', array_agg(post_reaction.author_id),
+                'count', count(post_reaction),
+                'content', post_reaction.content,
+                'emoji', (array_agg(emoji))[1]
+            )
+        FROM post_reaction
+        LEFT JOIN emoji
+        ON post_reaction.emoji_id = emoji.id
+        WHERE post_reaction.post_id = post.id
+            AND post_reaction.content IS NOT NULL
+        GROUP BY post_reaction.content
+    ) AS emoji_reactions";
 
 fn build_visibility_filter() -> String {
     format!(
@@ -550,7 +586,8 @@ pub async fn get_home_timeline(
             {related_mentions},
             {related_tags},
             {related_links},
-            {related_emojis}
+            {related_emojis},
+            {related_emoji_reactions}
         FROM post
         JOIN actor_profile ON post.author_id = actor_profile.id
         WHERE
@@ -620,6 +657,7 @@ pub async fn get_home_timeline(
         related_tags=RELATED_TAGS,
         related_links=RELATED_LINKS,
         related_emojis=RELATED_EMOJIS,
+        related_emoji_reactions=RELATED_EMOJI_REACTIONS,
         relationship_follow=i16::from(&RelationshipType::Follow),
         relationship_subscription=i16::from(&RelationshipType::Subscription),
         relationship_hide_reposts=i16::from(&RelationshipType::HideReposts),
@@ -660,7 +698,8 @@ pub async fn get_public_timeline(
             {related_mentions},
             {related_tags},
             {related_links},
-            {related_emojis}
+            {related_emojis},
+            {related_emoji_reactions}
         FROM post
         JOIN actor_profile ON post.author_id = actor_profile.id
         WHERE
@@ -677,6 +716,7 @@ pub async fn get_public_timeline(
         related_tags=RELATED_TAGS,
         related_links=RELATED_LINKS,
         related_emojis=RELATED_EMOJIS,
+        related_emoji_reactions=RELATED_EMOJI_REACTIONS,
         filter=filter,
         visibility_public=i16::from(&Visibility::Public),
         mute_filter=build_mute_filter(),
@@ -709,7 +749,8 @@ pub async fn get_direct_timeline(
             {related_mentions},
             {related_tags},
             {related_links},
-            {related_emojis}
+            {related_emojis},
+            {related_emoji_reactions}
         FROM post
         JOIN actor_profile ON post.author_id = actor_profile.id
         WHERE
@@ -731,6 +772,7 @@ pub async fn get_direct_timeline(
         related_tags=RELATED_TAGS,
         related_links=RELATED_LINKS,
         related_emojis=RELATED_EMOJIS,
+        related_emoji_reactions=RELATED_EMOJI_REACTIONS,
         visibility_direct=i16::from(&Visibility::Direct),
         mute_filter=build_mute_filter(),
     );
@@ -760,7 +802,8 @@ pub async fn get_related_posts(
             {related_mentions},
             {related_tags},
             {related_links},
-            {related_emojis}
+            {related_emojis},
+            {related_emoji_reactions}
         FROM post
         JOIN actor_profile ON post.author_id = actor_profile.id
         WHERE post.id IN (
@@ -787,6 +830,7 @@ pub async fn get_related_posts(
         related_tags=RELATED_TAGS,
         related_links=RELATED_LINKS,
         related_emojis=RELATED_EMOJIS,
+        related_emoji_reactions=RELATED_EMOJI_REACTIONS,
     );
     let rows = db_client.query(
         &statement,
@@ -839,7 +883,8 @@ pub async fn get_posts_by_author(
             {related_mentions},
             {related_tags},
             {related_links},
-            {related_emojis}
+            {related_emojis},
+            {related_emoji_reactions}
         FROM post
         JOIN actor_profile ON post.author_id = actor_profile.id
         WHERE {condition}
@@ -851,6 +896,7 @@ pub async fn get_posts_by_author(
         related_tags=RELATED_TAGS,
         related_links=RELATED_LINKS,
         related_emojis=RELATED_EMOJIS,
+        related_emoji_reactions=RELATED_EMOJI_REACTIONS,
         condition=condition,
     );
     let limit: i64 = limit.into();
@@ -884,7 +930,8 @@ pub async fn get_posts_by_tag(
             {related_mentions},
             {related_tags},
             {related_links},
-            {related_emojis}
+            {related_emojis},
+            {related_emoji_reactions}
         FROM post
         JOIN actor_profile ON post.author_id = actor_profile.id
         WHERE
@@ -903,6 +950,7 @@ pub async fn get_posts_by_tag(
         related_tags=RELATED_TAGS,
         related_links=RELATED_LINKS,
         related_emojis=RELATED_EMOJIS,
+        related_emoji_reactions=RELATED_EMOJI_REACTIONS,
         visibility_filter=build_visibility_filter(),
         mute_filter=build_mute_filter(),
     );
@@ -933,7 +981,8 @@ pub async fn get_post_by_id(
             {related_mentions},
             {related_tags},
             {related_links},
-            {related_emojis}
+            {related_emojis},
+            {related_emoji_reactions}
         FROM post
         JOIN actor_profile ON post.author_id = actor_profile.id
         WHERE post.id = $1
@@ -943,6 +992,7 @@ pub async fn get_post_by_id(
         related_tags=RELATED_TAGS,
         related_links=RELATED_LINKS,
         related_emojis=RELATED_EMOJIS,
+        related_emoji_reactions=RELATED_EMOJI_REACTIONS,
     );
     let maybe_row = db_client.query_opt(
         &statement,
@@ -988,7 +1038,8 @@ pub async fn get_thread(
             {related_mentions},
             {related_tags},
             {related_links},
-            {related_emojis}
+            {related_emojis},
+            {related_emoji_reactions}
         FROM post
         JOIN thread ON post.id = thread.id
         JOIN actor_profile ON post.author_id = actor_profile.id
@@ -1002,6 +1053,7 @@ pub async fn get_thread(
         related_tags=RELATED_TAGS,
         related_links=RELATED_LINKS,
         related_emojis=RELATED_EMOJIS,
+        related_emoji_reactions=RELATED_EMOJI_REACTIONS,
         visibility_filter=build_visibility_filter(),
         mute_filter=build_mute_filter(),
     );
@@ -1063,7 +1115,8 @@ pub async fn get_post_by_remote_object_id(
             {related_mentions},
             {related_tags},
             {related_links},
-            {related_emojis}
+            {related_emojis},
+            {related_emoji_reactions}
         FROM post
         JOIN actor_profile ON post.author_id = actor_profile.id
         WHERE post.object_id = $1
@@ -1073,6 +1126,7 @@ pub async fn get_post_by_remote_object_id(
         related_tags=RELATED_TAGS,
         related_links=RELATED_LINKS,
         related_emojis=RELATED_EMOJIS,
+        related_emoji_reactions=RELATED_EMOJI_REACTIONS,
     );
     let maybe_row = db_client.query_opt(
         &statement,
@@ -1440,7 +1494,8 @@ pub async fn search_posts(
             {related_mentions},
             {related_tags},
             {related_links},
-            {related_emojis}
+            {related_emojis},
+            {related_emoji_reactions}
         FROM post
         JOIN actor_profile ON post.author_id = actor_profile.id
         WHERE content_source ILIKE $1 AND author_id = $2
@@ -1452,6 +1507,7 @@ pub async fn search_posts(
         related_tags=RELATED_TAGS,
         related_links=RELATED_LINKS,
         related_emojis=RELATED_EMOJIS,
+        related_emoji_reactions=RELATED_EMOJI_REACTIONS,
     );
     let db_search_query = format!("%{}%", text);
     let rows = db_client.query(