+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Added `load-portable-object` command.
 - Perform proof verification if fetched object is portable.
+- Added `pleroma.emoji_reactions` list to `Status` entity.
 
 ### Changed